refactor(runtime-shared): rename providers/ → ports/ (ADR 0006)
Pure directory rename — no behaviour change, no class renames.
AesProviderBase, CredentialsProviderBase, and StorageProviderBase keep
their existing names; only the source path moves from src/providers/ to
src/ports/ to align vocabulary with ADR 0006 §1 (ports + adapters).

All nine import sites in the two host packages updated accordingly.
The adapters/ folders inside apps-host/electron and apps-host/web are
a follow-up task (the concrete adapter classes are untouched here).
