+++ apps-host/electron/src/lib/window-state-manager.ts
@@ -1,4 +1,4 @@
-import type { WindowState } from '@beak/runtime-shared/providers/storage';
+import type { WindowState } from '@beak/runtime-shared/ports/storage';
 import { type BrowserWindow, type BrowserWindowConstructorOptions, type Rectangle, screen } from 'electron';
 
 import getBeakHost from '../host';
